Separated classes into separate source files.
