+++ packml_sm/include/packml_sm/qt/cmd_transition.h
@@ -15,17 +15,12 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-
-#ifndef TRANSITIONS_H
-#define TRANSITIONS_H
+#pragma once
 
 #include "QEvent"
 #include "QAbstractTransition"
-
 #include "packml_sm/common.h"
 #include "packml_sm/state.h"
-
-#include "ros/console.h"
 
 namespace packml_sm
 {
@@ -36,38 +31,47 @@ public:
   {
     return new CmdTransition(CmdEnum::CLEAR, "clear", from, to);
   }
+
   static CmdTransition* start(PackmlState& from, PackmlState& to)
   {
     return new CmdTransition(CmdEnum::START, "start", from, to);
   }
+
   static CmdTransition* stop(PackmlState& from, PackmlState& to)
   {
     return new CmdTransition(CmdEnum::STOP, "stop", from, to);
   }
+
   static CmdTransition* hold(PackmlState& from, PackmlState& to)
   {
     return new CmdTransition(CmdEnum::HOLD, "hold", from, to);
   }
+
   static CmdTransition* abort(PackmlState& from, PackmlState& to)
   {
     return new CmdTransition(CmdEnum::ABORT, "abort", from, to);
   }
+
   static CmdTransition* reset(PackmlState& from, PackmlState& to)
   {
     return new CmdTransition(CmdEnum::RESET, "reset", from, to);
   }
+
   static CmdTransition* estop(PackmlState& from, PackmlState& to)
   {
     return new CmdTransition(CmdEnum::ESTOP, "estop", from, to);
   }
+
   static CmdTransition* suspend(PackmlState& from, PackmlState& to)
   {
     return new CmdTransition(CmdEnum::SUSPEND, "abort", from, to);
   }
+
   static CmdTransition* unsuspend(PackmlState& from, PackmlState& to)
   {
     return new CmdTransition(CmdEnum::UNSUSPEND, "unsuspend", from, to);
   }
+
   static CmdTransition* unhold(PackmlState& from, PackmlState& to)
   {
     return new CmdTransition(CmdEnum::UNHOLD, "unhold", from, to);
@@ -88,49 +92,4 @@ protected:
   CmdEnum cmd;
   QString name;
 };
-
-class StateCompleteTransition : public QAbstractTransition
-{
-public:
-  StateCompleteTransition()
-  {
-  }
-
-  StateCompleteTransition(PackmlState& from, PackmlState& to);
-
-  ~StateCompleteTransition()
-  {
-  }
-
-protected:
-  virtual bool eventTest(QEvent* e);
-  virtual void onTransition(QEvent* e)
-  {
-  }
-
-private:
-};
-
-class ErrorTransition : public QAbstractTransition
-{
-public:
-  ErrorTransition()
-  {
-  }
-
-  ErrorTransition(PackmlState& from, PackmlState& to);
-
-  ~ErrorTransition()
-  {
-  }
-
-protected:
-  virtual bool eventTest(QEvent* e);
-  virtual void onTransition(QEvent* e)
-  {
-  }
-
-private:
-};
 }
-#endif  // TRANSITIONS_H
