+++ packml_sm/include/packml_sm/state_machine_interface.h
@@ -16,6 +16,7 @@
  * limitations under the License.
  */
 #pragma once
+#include <functional>
 
 namespace packml_sm
 {
