+++ packml_sm/src/qt/state_machine.cpp
@@ -22,149 +22,6 @@
 
 namespace packml_sm
 {
-bool StateMachineInterface::start()
-{
-  switch (StatesEnum(getCurrentState()))
-  {
-    case StatesEnum::IDLE:
-      _start();
-      return true;
-    default:
-      ROS_WARN_STREAM("Ignoring START command in current state: " << getCurrentState());
-      return false;
-  }
-}
-
-bool StateMachineInterface::clear()
-{
-  switch (StatesEnum(getCurrentState()))
-  {
-    case StatesEnum::ABORTED:
-      _clear();
-      return true;
-    default:
-      ROS_WARN_STREAM("Ignoring CLEAR command in current state: " << getCurrentState());
-      return false;
-  }
-}
-
-bool StateMachineInterface::reset()
-{
-  switch (StatesEnum(getCurrentState()))
-  {
-    case StatesEnum::COMPLETE:
-    case StatesEnum::STOPPED:
-      _reset();
-      return true;
-    default:
-      ROS_WARN_STREAM("Ignoring RESET command in current state: " << getCurrentState());
-      return false;
-  }
-}
-
-bool StateMachineInterface::hold()
-{
-  switch (StatesEnum(getCurrentState()))
-  {
-    case StatesEnum::EXECUTE:
-      _hold();
-      return true;
-    default:
-      ROS_WARN_STREAM("Ignoring HOLD command in current state: " << getCurrentState());
-      return false;
-  }
-}
-
-bool StateMachineInterface::unhold()
-{
-  switch (StatesEnum(getCurrentState()))
-  {
-    case StatesEnum::HELD:
-      _unhold();
-      return true;
-    default:
-      ROS_WARN_STREAM("Ignoring HELD command in current state: " << getCurrentState());
-      return false;
-  }
-}
-
-bool StateMachineInterface::suspend()
-{
-  switch (StatesEnum(getCurrentState()))
-  {
-    case StatesEnum::EXECUTE:
-      _suspend();
-      return true;
-    default:
-      ROS_WARN_STREAM("Ignoring SUSPEND command in current state: " << getCurrentState());
-      return false;
-  }
-}
-
-bool StateMachineInterface::unsuspend()
-{
-  switch (StatesEnum(getCurrentState()))
-  {
-    case StatesEnum::SUSPENDED:
-      _unsuspend();
-      return true;
-    default:
-      ROS_WARN_STREAM("Ignoring UNSUSPEND command in current state: " << getCurrentState());
-      return false;
-  }
-}
-
-bool StateMachineInterface::stop()
-{
-  switch (StatesEnum(getCurrentState()))
-  {
-    case StatesEnum::STOPPABLE:
-    case StatesEnum::STARTING:
-    case StatesEnum::IDLE:
-    case StatesEnum::SUSPENDED:
-    case StatesEnum::EXECUTE:
-    case StatesEnum::HOLDING:
-    case StatesEnum::HELD:
-    case StatesEnum::SUSPENDING:
-    case StatesEnum::UNSUSPENDING:
-    case StatesEnum::UNHOLDING:
-    case StatesEnum::COMPLETING:
-    case StatesEnum::COMPLETE:
-      _stop();
-      return true;
-    default:
-      ROS_WARN_STREAM("Ignoring STOP command in current state: " << getCurrentState());
-      return false;
-  }
-}
-
-bool StateMachineInterface::abort()
-{
-  switch (StatesEnum(getCurrentState()))
-  {
-    case StatesEnum::ABORTABLE:
-    case StatesEnum::STOPPED:
-    case StatesEnum::STARTING:
-    case StatesEnum::IDLE:
-    case StatesEnum::SUSPENDED:
-    case StatesEnum::EXECUTE:
-    case StatesEnum::HOLDING:
-    case StatesEnum::HELD:
-    case StatesEnum::SUSPENDING:
-    case StatesEnum::UNSUSPENDING:
-    case StatesEnum::UNHOLDING:
-    case StatesEnum::COMPLETING:
-    case StatesEnum::COMPLETE:
-    case StatesEnum::CLEARING:
-    case StatesEnum::STOPPING:
-      _abort();
-      return true;
-    default:
-      ROS_WARN_STREAM("Ignoring ABORT command in current state: " << getCurrentState());
-      return false;
-  }
-}
-
 QCoreApplication* a;
 void init(int argc, char* argv[])
 {
@@ -410,73 +267,4 @@ void StateMachine::_abort()
 {
   sm_internal_.postEvent(CmdEvent::abort());
 }
-
-ContinuousCycle::ContinuousCycle()
-{
-  ROS_INFO_STREAM("Forming CONTINUOUS CYCLE state machine (states + transitions)");
-
-  // Naming <from state>_<to state>
-  CmdTransition* abortable_aborting_on_cmd = CmdTransition::abort(*abortable_, *aborting_);
-  ErrorTransition* abortable_aborting_on_error = new ErrorTransition(*abortable_, *aborting_);
-  StateCompleteTransition* aborting_aborted = new StateCompleteTransition(*aborting_, *aborted_);
-  CmdTransition* aborted_clearing_ = CmdTransition::clear(*aborted_, *clearing_);
-  StateCompleteTransition* clearing_stopped_ = new StateCompleteTransition(*clearing_, *stopped_);
-  CmdTransition* stoppable_stopping_ = CmdTransition::stop(*stoppable_, *stopping_);
-  StateCompleteTransition* stopping_stopped = new StateCompleteTransition(*stopping_, *stopped_);
-  CmdTransition* stopped_resetting_ = CmdTransition::reset(*stopped_, *resetting_);
-  StateCompleteTransition* unholding_execute_ = new StateCompleteTransition(*unholding_, *execute_);
-  CmdTransition* held_unholding_ = CmdTransition::unhold(*held_, *unholding_);
-  StateCompleteTransition* holding_held_ = new StateCompleteTransition(*holding_, *held_);
-  CmdTransition* idle_starting_ = CmdTransition::start(*idle_, *starting_);
-  StateCompleteTransition* starting_execute_ = new StateCompleteTransition(*starting_, *execute_);
-  CmdTransition* execute_holding_ = CmdTransition::hold(*execute_, *holding_);
-  StateCompleteTransition* execute_execute_ = new StateCompleteTransition(*execute_, *execute_);
-  StateCompleteTransition* completing_complete = new StateCompleteTransition(*completing_, *complete_);
-  CmdTransition* complete_resetting_ = CmdTransition::reset(*complete_, *resetting_);
-  StateCompleteTransition* resetting_idle_ = new StateCompleteTransition(*resetting_, *idle_);
-  CmdTransition* execute_suspending_ = CmdTransition::suspend(*execute_, *suspending_);
-  StateCompleteTransition* suspending_suspended_ = new StateCompleteTransition(*suspending_, *suspended_);
-  CmdTransition* suspended_unsuspending_ = CmdTransition::unsuspend(*suspended_, *unsuspending_);
-  StateCompleteTransition* unsuspending_execute_ = new StateCompleteTransition(*unsuspending_, *execute_);
-
-  abortable_->setInitialState(clearing_);
-  stoppable_->setInitialState(resetting_);
-  sm_internal_.setInitialState(aborted_);
-  ROS_INFO_STREAM("State machine formed");
-}
-
-SingleCycle::SingleCycle()
-{
-  ROS_INFO_STREAM("Forming SINGLE CYCLE state machine (states + transitions)");
-
-  // Naming <from state>_<to state>
-  CmdTransition* abortable_aborting_on_cmd = CmdTransition::abort(*abortable_, *aborting_);
-  ErrorTransition* abortable_aborting_on_error = new ErrorTransition(*abortable_, *aborting_);
-  StateCompleteTransition* aborting_aborted = new StateCompleteTransition(*aborting_, *aborted_);
-  CmdTransition* aborted_clearing_ = CmdTransition::clear(*aborted_, *clearing_);
-  StateCompleteTransition* clearing_stopped_ = new StateCompleteTransition(*clearing_, *stopped_);
-  CmdTransition* stoppable_stopping_ = CmdTransition::stop(*stoppable_, *stopping_);
-  StateCompleteTransition* stopping_stopped = new StateCompleteTransition(*stopping_, *stopped_);
-  CmdTransition* stopped_resetting_ = CmdTransition::reset(*stopped_, *resetting_);
-  StateCompleteTransition* unholding_execute_ = new StateCompleteTransition(*unholding_, *execute_);
-  CmdTransition* held_unholding_ = CmdTransition::unhold(*held_, *unholding_);
-  StateCompleteTransition* holding_held_ = new StateCompleteTransition(*holding_, *held_);
-  CmdTransition* idle_starting_ = CmdTransition::start(*idle_, *starting_);
-  StateCompleteTransition* starting_execute_ = new StateCompleteTransition(*starting_, *execute_);
-  CmdTransition* execute_holding_ = CmdTransition::hold(*execute_, *holding_);
-  StateCompleteTransition* execute_completing_ = new StateCompleteTransition(*execute_, *completing_);
-  StateCompleteTransition* completing_complete = new StateCompleteTransition(*completing_, *complete_);
-  CmdTransition* complete_resetting_ = CmdTransition::reset(*complete_, *resetting_);
-  StateCompleteTransition* resetting_idle_ = new StateCompleteTransition(*resetting_, *idle_);
-  CmdTransition* execute_suspending_ = CmdTransition::suspend(*execute_, *suspending_);
-  StateCompleteTransition* suspending_suspended_ = new StateCompleteTransition(*suspending_, *suspended_);
-  CmdTransition* suspended_unsuspending_ = CmdTransition::unsuspend(*suspended_, *unsuspending_);
-  StateCompleteTransition* unsuspending_execute_ = new StateCompleteTransition(*unsuspending_, *execute_);
-
-  abortable_->setInitialState(clearing_);
-  stoppable_->setInitialState(resetting_);
-  sm_internal_.setInitialState(aborted_);
-  ROS_INFO_STREAM("State machine formed");
-}
-
 }  // packml_sm
